+++ cmake/travis_script_linux.sh
@@ -7,13 +7,9 @@ if [ "$CC" = "clang" ]; then export CC="clang-3.5"; fi
 if [ "$CXX" = "clang++" ]; then export CXX="clang++-3.5"; fi
 
 # Need to build latest glbinding from source
-#GLBVER=`curl https://api.github.com/repos/hpicgs/glbinding/releases/latest | grep -m1 tag_name | cut -d\" -f4`
-#curl -L https://github.com/hpicgs/glbinding/archive/${GLBVER}.tar.gz | tar xzf -
-#cd glbinding-${GLBVER#v}
-
-# Need to use master until a fixed version is released
-git clone https://github.com/hpicgs/glbinding.git
-cd glbinding
+GLBVER=`curl https://api.github.com/repos/hpicgs/glbinding/releases/latest | grep -m1 tag_name | cut -d\" -f4`
+curl -L https://github.com/hpicgs/glbinding/archive/${GLBVER}.tar.gz | tar xzf -
+cd glbinding-${GLBVER#v}
 
 mkdir build
 cd build