+++ cmake/travis_script_mac.sh
@@ -1,13 +1,9 @@
 #!/bin/bash
 
 # Need to build latest glbinding from source
-#GLBVER=`curl https://api.github.com/repos/hpicgs/glbinding/releases/latest | grep -m1 tag_name | cut -d\" -f4`
-#curl -L https://github.com/hpicgs/glbinding/archive/${GLBVER}.tar.gz | tar xzf -
-#cd glbinding-${GLBVER#v}
-
-# Need to use master until a fixed version is released
-git clone https://github.com/hpicgs/glbinding.git
-cd glbinding
+GLBVER=`curl https://api.github.com/repos/hpicgs/glbinding/releases/latest | grep -m1 tag_name | cut -d\" -f4`
+curl -L https://github.com/hpicgs/glbinding/archive/${GLBVER}.tar.gz | tar xzf -
+cd glbinding-${GLBVER#v}
 
 mkdir build
 cd build